refactor(score-analytics): extract ClickHouse query building logic
Extract query building functions from scoreAnalyticsRouter.ts into separate
files for better maintainability and code organization:

- buildEstimateQuery.ts: Preflight estimation query with 1% sampling
- buildScoreComparisonQuery.ts: Main analytics query with ~1000 lines of
  CTE-based query logic

The main query remains as a single cohesive CTE chain to preserve
dependencies between filtered datasets, bounds, and analytics CTEs.
Helper functions for filters and sampling are included as internal
utilities within each query builder.

Router reduced from ~1,600 lines to ~400 lines while maintaining full
functionality and test coverage.

Note: Query performance characteristics remain unchanged. Future
consideration for splitting CTEs into separate queries documented in
buildScoreComparisonQuery.ts for gradual migration to centralized
metric query builder interface.
